Traktor: OpenGL renderer fixes.
[git-p4: depot-paths = "//depot/traktor/": change = 256]

+++ code/Render/OpenGL/RenderTargetOpenGL.cpp
@@ -173,12 +173,14 @@ void RenderTargetOpenGL::destroy()
 {
 	if (m_colorTexture)
 	{
-		m_context->deleteResource(new DeleteTextureCallback(m_colorTexture));
+		if (m_context)
+			m_context->deleteResource(new DeleteTextureCallback(m_colorTexture));
 		m_colorTexture = 0;
 	}
 	if (m_frameBufferObject)
 	{
-		m_context->deleteResource(new DeleteFramebufferCallback(m_frameBufferObject));
+		if (m_context)
+			m_context->deleteResource(new DeleteFramebufferCallback(m_frameBufferObject));
 		m_frameBufferObject = 0;
 	}
 }
